refactor(#92): move test doubles to the test files.
