+++ pkg/server/config/exporters/exporters_test.go
@@ -115,3 +115,23 @@ func TestExportToYAML_ShouldWriteFile_WhenConfigIsValid(t *testing.T) {
 	require.Contains(t, string(data), "a: default_hello")
 	require.Contains(t, string(data), "d_nested_field: default_world")
 }
+
+type ExporterTestConfig struct {
+	A string                `mapstructure:"a"`
+	B int                   `mapstructure:"b_with_long_name"`
+	C ExporterTestSubConfig `mapstructure:"c_sub_config"`
+}
+
+type ExporterTestSubConfig struct {
+	D string `mapstructure:"d_nested_field"`
+}
+
+func NewExporterTestConfig() ExporterTestConfig {
+	return ExporterTestConfig{
+		A: "default_hello",
+		B: 1,
+		C: ExporterTestSubConfig{
+			D: "default_world",
+		},
+	}
+}
